requirements.txt
pyserial
pysimplegui